Placed data analysis methods in seperate file, added code to calculate and visualize dissimilarity between digit classes
